requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.1
whitenoise==6.6.0
requests==2.31.0
gunicorn==21.2.0
